requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg[binary]==3.1.18
python-dotenv==1.0.1
stripe==8.10.0
pydantic==2.8.2
python-dateutil==2.9.0.post0

